requirements.txt
# Core dependencies
biopython>=1.79
click>=8.0.0
pyyaml>=6.0
psutil>=5.8.0

# Optional scientific computing dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and testing dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991